v2: split domains into separate files
Updates tailscale/corp#21867
