+++ v2/dns_test.go
@@ -1,0 +1,182 @@
+package tailscale_test
+
+import (
+	"context"
+	"encoding/json"
+	"net/http"
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/tailscale/tailscale-client-go/v2"
+)
+
+func TestClient_DNSNameservers(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	expectedNameservers := map[string][]string{
+		"dns": {"127.0.0.1"},
+	}
+
+	server.ResponseBody = expectedNameservers
+	nameservers, err := client.DNSNameservers(context.Background())
+	assert.NoError(t, err)
+	assert.Equal(t, http.MethodGet, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/nameservers", server.Path)
+	assert.Equal(t, expectedNameservers["dns"], nameservers)
+}
+
+func TestClient_DNSPreferences(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+	server.ResponseBody = &tailscale.DNSPreferences{
+		MagicDNS: true,
+	}
+
+	preferences, err := client.DNSPreferences(context.Background())
+	assert.NoError(t, err)
+	assert.Equal(t, http.MethodGet, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/preferences", server.Path)
+	assert.Equal(t, server.ResponseBody, preferences)
+}
+
+func TestClient_DNSSearchPaths(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	expectedPaths := map[string][]string{
+		"searchPaths": {"test"},
+	}
+
+	server.ResponseBody = expectedPaths
+
+	paths, err := client.DNSSearchPaths(context.Background())
+	assert.NoError(t, err)
+	assert.Equal(t, http.MethodGet, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/searchpaths", server.Path)
+	assert.Equal(t, expectedPaths["searchPaths"], paths)
+}
+
+func TestClient_SplitDNS(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	expectedNameservers := tailscale.SplitDnsResponse{
+		"example.com": {"1.1.1.1", "1.2.3.4"},
+	}
+
+	server.ResponseBody = expectedNameservers
+	nameservers, err := client.SplitDNS(context.Background())
+	assert.NoError(t, err)
+	assert.Equal(t, http.MethodGet, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/split-dns", server.Path)
+	assert.Equal(t, expectedNameservers, nameservers)
+}
+
+func TestClient_SetDNSNameservers(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	nameservers := []string{"127.0.0.1"}
+
+	assert.NoError(t, client.SetDNSNameservers(context.Background(), nameservers))
+	assert.Equal(t, http.MethodPost, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/nameservers", server.Path)
+
+	body := make(map[string][]string)
+	assert.NoError(t, json.Unmarshal(server.Body.Bytes(), &body))
+	assert.EqualValues(t, nameservers, body["dns"])
+}
+
+func TestClient_SetDNSPreferences(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	preferences := tailscale.DNSPreferences{
+		MagicDNS: true,
+	}
+
+	assert.NoError(t, client.SetDNSPreferences(context.Background(), preferences))
+	assert.Equal(t, http.MethodPost, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/preferences", server.Path)
+
+	var body tailscale.DNSPreferences
+	assert.NoError(t, json.Unmarshal(server.Body.Bytes(), &body))
+	assert.EqualValues(t, preferences, body)
+}
+
+func TestClient_SetDNSSearchPaths(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	paths := []string{"test"}
+
+	assert.NoError(t, client.SetDNSSearchPaths(context.Background(), paths))
+	assert.Equal(t, http.MethodPost, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/searchpaths", server.Path)
+
+	body := make(map[string][]string)
+	assert.NoError(t, json.Unmarshal(server.Body.Bytes(), &body))
+	assert.EqualValues(t, paths, body["searchPaths"])
+}
+
+func TestClient_UpdateSplitDNS(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	nameservers := []string{"1.1.2.1", "3.3.3.4"}
+	request := tailscale.SplitDnsRequest{
+		"example.com": nameservers,
+	}
+
+	expectedNameservers := tailscale.SplitDnsResponse{
+		"example.com": nameservers,
+	}
+	server.ResponseBody = expectedNameservers
+
+	resp, err := client.UpdateSplitDNS(context.Background(), request)
+	assert.NoError(t, err)
+	assert.Equal(t, http.MethodPatch, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/split-dns", server.Path)
+
+	body := make(tailscale.SplitDnsResponse)
+	assert.NoError(t, json.Unmarshal(server.Body.Bytes(), &body))
+	assert.EqualValues(t, nameservers, body["example.com"])
+	assert.Equal(t, expectedNameservers, resp)
+}
+
+func TestClient_SetSplitDNS(t *testing.T) {
+	t.Parallel()
+
+	client, server := NewTestHarness(t)
+	server.ResponseCode = http.StatusOK
+
+	nameservers := []string{"1.1.2.1", "3.3.3.4"}
+	request := tailscale.SplitDnsRequest{
+		"example.com": nameservers,
+	}
+
+	assert.NoError(t, client.SetSplitDNS(context.Background(), request))
+	assert.Equal(t, http.MethodPut, server.Method)
+	assert.Equal(t, "/api/v2/tailnet/example.com/dns/split-dns", server.Path)
+
+	body := make(tailscale.SplitDnsResponse)
+	assert.NoError(t, json.Unmarshal(server.Body.Bytes(), &body))
+	assert.EqualValues(t, nameservers, body["example.com"])
+}
